Conversion: Add option to embed all referenced fonts
Conversion: Add an option to embed all fonts that are referenced in the
input document but are not already embedded. This will search your
system for the referenced font, and if found, the font will be embedded.
Only works if the output format supports font embedding (for example:
EPUB or AZW3).