Bulk metadata edit: option to not refresh after edit
Bulk metadata edit: Add a checkbox to prevent the refreshing of the book
list after the bulk edit. This means that the book list will not be
resorted and any existing search/virtual library will not be refreshed.
Useful if you have a large library as the refresh can be slow.