Append and modify CSS when sending epubs to Kobo devices
This allows users to have a "kobo_extra.css" file in the root of their
device containing CSS rules. This will be appended to all stylesheets in
the epub. As well, if the extra rules contain an @page rule, any
existing @page rules will be stripped from the stylesheets. Finally, if
any of the extra rules include "widows" and "orphans" settings, these
are stripped from the rules in the stylesheets.