Preserve current book when doing searches
When doing searches or switching between virtual libraries in the main
book list, preserve the current book. The currently selected book will
remain visible if it is present in the results of the search or the
selected virtual library. Fixes #1216713 [After selecting books which belongs to same serie, when returning to main window, we are at the top of the list and not at previously selected item](https://bugs.launchpad.net/calibre/+bug/1216713)