Reading status wiped for older Kobo Touch firmware
Kobo driver: Fix reading status being cleared when connecting to a Kobo
with older firmware and metadata management set to automatic. Fixes #1230018 [Reading status wiped for older Kobo Firmware](https://bugs.launchpad.net/calibre/+bug/1230018)

When a Kobo Touch running a 1.9.x firmware was connected, the reading
status of all books was wiped out. This was to do with the way the
collections were managed. If a shelf/collection attribute was not
defined, the reading status was being cleared when it should not have
been.

Also skipping a loop over the books if shelf and series management is
not being done.