Fix bug in handling change of filename but not foldername
Fix a bug triggered in very special circumstances (primarily on windows)
that could cause duplicate files entries (hard links) to be created for
a books' files when changing the title and author. The bug was triggered
only if the title or author were between 32 and 35 characters in length
and the book entry had originally been created by a pre 1.x version of
calibre.