Conversion: When converting TXT set <title> in HTML
TXT Input: Ensure that <title> in the generated HTML has a
meaningful value. Fixes #1236923 [txt → epub When I don't give a title, the <title> tag is set to unknown. It should default to the chapter name and chapter ID for each fragment. Or use the name of the converted file](https://bugs.launchpad.net/calibre/+bug/1236923)