Added a device option: opt_do_not_sanitize_filename
This option just enables to not sanitize a filename of the content at all.
Some devices, such as Amazon Kindle, for PDF books, use a filename as a title to display.
In non-ascii encoding region, filenames may contains not only ascii
characters but also multibyte characters, which are currently sanitized
to only-ascii characters.

This must be inconvenience to the users of multibyte characters.

So I added opt_do_not_sanitize_filename to DeviceConfig,
and at the first, enabled the option for Kinde2(and the later)
because I can confirm that device.