Update profiles.py
Add profile for Galaxy Tab 3 8, i.e. screen size and dpi.

Explanation. I read exclusively PDFs all platforms, because the annotation facilities are extensive and portable across platforms, which unfortunately can't be said for any of the other ebook formats I know of. Given that PDFs don't reflow, I need to know the screensize of the device I'm generating for, in my case a Samsung Tab 3 8 tablet. I hope it will be useful to others.