Update to support Kobo firmware 3.4.1
Kobo are about to release firmware 3.4.1. The beta and release
candidates have been tested with calibre and the version number is all
that needs to change. As usual, I am bumping it to 3.4.2 in case of a
second bug fix release with no backend changes. Fixes #1324375 [Update to support Kobo firmware 3.4.1](https://bugs.launchpad.net/calibre/+bug/1324375)