Edit Book: Run syntax highlighting in the background so that it does not affect responsiveness of the user interface
Technically, the syntax highlighting now does idle processing, that is
it yields after every line of text is highlighted to allow the rest of
the UI to respond to user interaction. This gives an excellent
illusion of running int he background, except for pathological files
with very long lines (such as those that have all their markup on a
single line).