Various improvements to the splash screen
On windows and linux draw the splash screen in a worker process so that
the splash screen is updated properly despite the UI thread being
blocked while calibre is starting up. Fixes #1357553 [Splash screen is messed up in calibre 2.05 beta](https://bugs.launchpad.net/calibre/+bug/1357553)

Also, draw the messages in the splash screen on an opaque beackground so
that  they can always be read.

Drawing the splash screen in a separate process does not work on OS X so
we continue to use the old technique there.