New option to override page count calculation method on book-by-book basis
A new option is provided which allows the user to create a custom column which specifies which apnx calculation method to use for a particular book.  The original method option is treated as a default which is overridden in cases where a valid method is given in the specified column.  This is useful, for example, in cases where the user has a small number of books which are setup for the pagebreak method and wants to use the accurate method most of the rest of the time.

Because this option was introduced into the option list before the overwrite option, I've also included code which should shift the options around accordingly in cases where the overwrite option has been implemented and used prior to applying this change.