Download metadata dialog: Get rid of the Next button
I am tired of being asked why the Next button and the OK button do the
same thing, even though they dont, quite. Since the purpose of the Next
button is to prevent the Back button from changing button positions,
when it appears, remove the Next button and move the Back button out of
the button box, so that it cannot change button positions.