Load system tray icon from the current theme
Loading the tray icon from the theme allows the user to customize the
desktop in greater extent. If the icon is not present in the icon set,
then the bundled one is loaded, so the previous behavior is preserved.
The themed icon name is consistent with the application name -
calibre-gui.