Don't block UI when testing email with wizard
The test_email dialog in the wizard runs the email test in the same
thread as the UI. In the case where there's a timeout trying to send the
email, this can block the UI for up to thirty seconds.

This patch runs the email test in a new thread, and updates the
QTPlainTextEdit field via a slot rather than manually.