Speedup reading metadata from OPF files by not reading the NCX
Also do not render the first page of the EPUB if it is encrypted. The
old logic for detecting this was broken in the case where the OPF file
is not in the root of the EPUB. Also make the OPF metadata reading code
no longer depend directly on the opf2.OPF class.