Add device Aura One and support for ignoring some collecitons.
Kobo are releasing the Aura One and Aura Edition 2. This adds the device
ids but does not update the firmware version. The firmware for these
devices have not been released. It will probably be compatible, but,
there might be changes.

This also includes an extra configuration option to ignore some
collections. With this, the named collections will not be touched buy
the drivers collection management.