Rework cover upload
Reworked the cover image generation for uploading. There was an error in
calculating the size, the Glo HD sizes needed to be diffrent, the cover
file was being read multiple times and the full size was not being
resized when it should have. Also, some refactoring was done at the
request of someone subclassing the driver.

Lastly, the firmware build version has been bumped to match the latest
release from Kobo.