Adds somewhat graceful handling for Ebook split errors.
It is difficult to reproduce exactly, but some feeds that I have used
will cause a SplitError.  This will occur no matter which preferences
are changed, such as split point, disabling split in the UI, etc.

So, this acts as a last resort to prevent the split error from halting
all processing of the ebook.