localization: differentiate between C and not set Linux env. variables
In case of e.g. LANG=hu_HU.UTF-8 and LC_MESSAGES=C, the expected
behavior is that the UI is English, not Hungarian.

The problem was that we did not differentiate between a not set
LC_MESSAGES and one that is set to C, as locale.getdefaultlocale()
returns None in both cases.