New option to override KoboTouch driver book resend behaviour and bump the supported firmware version
Adding an option for overriding the Kobo firmware behaviour when a book
is replaced. The defualt behaviour is for the book to be removed from
the internal database and then treated as a new book. This means the
reading status is lost. But, the KoboTouch driver has always overriden
this behaviour and to allow the reading status to be kept. The option is
to override the firmware behaviour and defaults to on to keep the
existing driver behaviour.

Kobo has released new firmware. The only change is to bump the firmware
and database versions.