Add comments to metadata plugboard
This allows modification of comments to add info not displayed on the
device as part of the comments/description/synopsis or whatever on the
device. It was discussed in
https://www.mobileread.com/forums/showthread.php?t=291138. I have been
running this since before then with no issues for Kobo devices and
save-to-disk. It includes a warning prompt as mangled comments could
cause a problem. But, mangled any field could as well.