Devendor markdown.
In order to retain compatibility with the current API, markdown is still
exported as calibre.ebooks.markdown

In order to ensure gracefully handling the list of extensions changing
from version to version, filter them based on what is detected as
present at runtime. This is necessary because the headerid extension did
in fact get removed from python-markdown 3.0

Unfortunately, markdown itself does not have a good way to get names +
descriptions, so this cannot be detected in a fully dynamic manner.