py3: also make plugin installer work
We check if versions compare favorably to the installed version, and if
it None, then we report that it can be installed. But this check always
returns None < available_version on python2, and on python3 it is a
comparison error -- so lift out the None handling to return as early as
possible.