py3: port oeb rasterize string types
The problem is I'm not sure what manifest's item.data is supposed to be.
The docs for calibre.ebooks.oeb.base.Manifest.Item.data say "it should
be a str type", but that doesn't seem to make sense for a binary file
(at least for python3). More likely it should be a string_or_bytes, and
consumers should check whether the item data is an instance of bytes
(which it should be for image files that aren't representable as decoded
bytes).