py3: write unicode bom without using bytes() type
Since the file is opened in utf-8 mode, it needs to be written to using
unicode text, not a b'' string. In python3, '\xef\xbb\xbf' becomes 'ï»¿'
which is definitely not a BOM. Writing the unicode escaped codepoint
allows the codecs.open() encoding to correctly choose the right bytes
for the BOM and insert it as needed.