debug: return the fully qualified path to calibre-debug
We need to use sys.executables_location, because there may be multiple
calibre-debug programs in the path or we may simply not be in the path
at all. I ran into this due to testing multiple installations of calibre
for python2/python3, and discovering that run_calibre_debug used the
wrong python version.