CFI Feedback Fixes, Paged Mode Crash
Use original getClientRects instead of getBoundingClientRect in range_has_point. getBoundingClientRect will not work in some cases.

In cfi.pyj's scroll_to, store the range's node's parent node and the index of the range's node in the parent and use that to restore the range after merging the span back in.

Also, don't store the span's bounding rect, since it's not needed anymore.

Fixed a typo in snap_to_selection in paged_mode that would cause a crash.