Use python 3.7 compatible syntax.
The := operator was introduced in python 3.8. I don't think making calibre incompatible with previous python versions is worth the use of the new operator here.

FreeBSD at least uses python 3.7 by default and this would cause problems in providing calibre to users, and require them to custom compile software to get it.