Bug #1948560: Odd behaviour with digits in alphabetical partitioning
This is fixed for the tag browser in the gui and the content server. I couldn't find anywhere else that used first letter partitioning.

When using first letter partitioning get_categories() constructs the sort key as follows:
- If the first character of the string is a digit (I hope isdigit() is localized) then use that digit otherwise use a string that is larger than any digit
- The ICU collation order key
- The sort key.

As such, items that start with a digit will first sort by that digit, then if needed use the ICU key as the second level then if needed use the sort key as the third level. Items that do not start with a digit will sort by the ICU key then the sort key.