fix the linux desktop filename mapping for ebook-edit
This was initially added in commit
efb83eb6fcd1b175e70281fc0776ca2e53422ad0 with a filler name that,
however, didn't match the usual style of desktop filenames. When the
actual desktop file was added in commit
a87092ba4f5e61aa51149c2c34880582613037f0 it didn't match the filename.

Although it was mass renamed from "tweak" to "edit" in commit
289ef5f0b9a4a87e6918744af64ff0d80f6a6c42, the word ordering was still
wrong.

As a result, application menus failed to correlate open windows to
pinned icons, or (on application menus that support it) group the
windows together.