Remove some repeated links to other articles
On the web, under some articles, there are links to other articles or sections. These links are not present in the print edition and are repetetive and annoying. For example, these links are something like this: "For exclusive insight and reading recommendations from our correspondents in America, sign up to Checks and Balance, our weekly newsletter."

These links follow the ■ symbol.

This commit removes them and anything that follows the ■ symbol marking the end of the article (so they are safe to remove).

The only downside is that applying the ■ symbol is not consistent - some articles do not end by it. One could target specific texts to remove as they are repetetive (like the one quoted above that appears under articles in the US sections), but that sounds maintanance-intensive.