Fix a regression that broke handling of mouse clicks on some Linux Wayland systems with a touchpad
Basically, in Qt 6 Wayland, Qt reports the device for mouse events
as finger when they come from a touchpad instead of a mouse. This caused
them to be swallowed by the gesture recognizer as being synthesized
mouse events.

Fixes #1981629 [Viewer TOC and search results  are not clickable](https://bugs.launchpad.net/calibre/+bug/1981629)