Get rid of BasicSplitter
Move custom drawing of splitter handle into the style where it belongs.
This means all splitters now use the custom handle without needing to
sprinkle a special class everywhere.

Note that this only works when using the calibre style not the system
one. This is by design, since we do not want to override drawing of
"system" components.