Try getting NYTimes from the internet archive
nytimes.com is apparently doing a lot of bot detection. Let's see if
getting it from the Internet Archive works better. This goes via a
caching proxy so that any article is only downloaded once from the
archive regardless of how many users download using the recipe. This
ensures the archive is not unduly loaded.