PDF Output: Fix background image + text not rendering correctly if the same background image is used with different text multiple times. Fixes #2035338 [conversion from epub to pdf - all chapter titles are the same as the first chapter](https://bugs.launchpad.net/calibre/+bug/2035338)
Chromium renders this as the background image with a soft mask. The
background image is duplicated, however calibre's image dedup
cannot work as the softmask is part of the image dictionary and dedup
works by replacing references to the image dictionary.

There is a similar breakage which will happen if there are duplicated
soft masks as the dedup code does not currently update the softmask
references.