Change to using a fixed standard light palette
This fixes all manner of subtle issues introduced by Qt's broken
attempts to map system colors to a QPalette.

1) Highlight color being same as Windows' accent color (this is anyway
   reverted in Qt 6.6)

2) Window titlebars not following system light/dark changes on Windows.

3) All the custom code that I wrote for tracking system light/dark
   changes can be dropped in favor of the new QStyleHints::colorScheme
   property

4) We had a fixed dark mode palette but not a fixed light mode palette.

I am sure people, especially on Linux, are going to complain about this,
but there are simply too many upsides. I will perhaps add a way to
configure the details of the light and dark palettes for people that
need it. This dialog could theoretically have a button to try to import
the standard system palette automatically to create the custom palette.