tests: refactor out speech-related Qt tests to distinct testcase
TTS support can be gated on e.g. Gentoo's USE=speech, so we don't want
to run the basic selfcheck of speech support if this is disabled. We do
want to run the general Qt checks, though (and also do a selfcheck of
speech support if the USE=speech is enabled!) -- this is much more
conveniently done if there is a distinct testcase for it that can be
passed into --exclude-test-name.