Make the attributes id and has_cover 'official' calibre metadata attributes. This allows them to be copied by deepcopy_metadata().
The problem was exposed by a kiwidude plugin "Search the Internet". After a deepcopy_metadata() a template using "{id}" failed with "unknown field".

I don't think this change will break anything. If you are worried then reject the PR. The problem has been there since forever, so it clearly is not a burning problem.