PDF Input: Automatic header/footer detection and removal
A new, dedicated PDF Input engine for calibre that implements
automatic detection of headers and footers based on document analysis.
The new engine is the default to go back to using the old engine, select
it in the PDF Input section of the Conversion dialog. See #2076346 (PDF conversion new engine does not work)