Use stable ordering when writing shell completions
Set ordering is unstable (uses the hash for ordering), so the output file
would be different every time. Sort the sets into lists before writing. In
various places there were unnecessary "conversions", e.g. sorted() always
returns a list, so 'list(sorted(…))' is not useful.

In one place output is fixed to say: "Choices are: doc, ereader, ztxt" instead
of "['doc', 'ereader', 'ztxt']".

For module constants, just create the object directly. The code is easier to
read/edit when the list is constructed directly with items that are ordered in
the same way that is shown in various outputs (instead of being sorted
dynamically).

There is some more syntax confusion in src/calibre/linux.py. For example,
'"%s" % (x)' is the same as '"%s" % x', because just putting something in
parentheses does not create a tuple. It'd be nice to clean up this whole file
to use f-strings and not the error-prone %-formatting, but that's a separate
issue.

In a test build of the Fedora package, the generated file is now stable.