constants: drop DeVendor for cryptodome
Converting Cryptodome imports to Crypto will
fall back to old Crypto library. For code that uses
Cryptodome syntax it will fail when using function
names that does not exist.

For example if python-crypto is not installed,
then error will be
    	from Crypto import Random
	ModuleNotFoundError: No module named 'Crypto'

Or if python-crypto is installed, but calling incompatible
functions, getting error
	AttributeError: module 'Crypto.PublicKey.RSA' has no attribute 'import_key'. Did you mean: 'importKey'?

Signed-off-by: Tomas Melin <tomas.melin@iki.fi>